Break down Early_Access into smaller classes
